Remove some duplicated code for filtering attempts
