+++ lib/stats/charts/attempts_by_time.dart
@@ -1,8 +1,12 @@
+import 'dart:async';
+
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:sendrax/models/attempt.dart';
 import 'package:sendrax/util/constants.dart';
+
+import 'attempt_filter.dart';
 
 class AttemptsByTimeChart extends StatefulWidget {
   AttemptsByTimeChart({Key key, @required this.attempts, @required this.locationNamesToIds})
@@ -15,19 +19,15 @@ class AttemptsByTimeChart extends StatefulWidget {
 }
 
 class _AttemptsByTimeChartState extends State<AttemptsByTimeChart> {
-  List<charts.Series> chartSeries;
   List<charts.TickSpec<String>> ticks;
-  String filterTimeframe;
-  String filterLocation;
-  String filterSendType;
-  String filterCategory;
+  StreamController<List<Attempt>> filteredAttemptsStream;
 
   int selectedHour;
   int selectedCount;
 
   @override
   void initState() {
-    chartSeries = _buildChartSeries(context);
+    filteredAttemptsStream = StreamController<List<Attempt>>.broadcast();
     ticks = _buildTicks();
     super.initState();
   }
@@ -35,7 +35,10 @@ class _AttemptsByTimeChartState extends State<AttemptsByTimeChart> {
   @override
   Widget build(BuildContext context) {
     List<Widget> children = <Widget>[];
-    children.add(_buildFilters(context));
+    children.add(AttemptFilter(
+        attempts: widget.attempts,
+        locationNamesToIds: widget.locationNamesToIds,
+        filteredAttemptsStream: filteredAttemptsStream));
     children.add(_buildChart(context));
     DateTime time = DateTime.now();
     if (selectedHour != null) {
@@ -68,204 +71,7 @@ class _AttemptsByTimeChartState extends State<AttemptsByTimeChart> {
     );
   }
 
-  Widget _buildFilters(BuildContext context) {
-    return Container(
-        child: Row(children: <Widget>[
-      Expanded(
-          child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          Row(
-            children: <Widget>[
-              _showTimeFrameDropdown(context),
-              _showLocationDropdown(context),
-            ],
-          ),
-          Row(
-            children: <Widget>[
-              _showSendTypeDropdown(context),
-              _showCategoryDropdown(context),
-            ],
-          )
-        ],
-      )),
-      _showClearDropdownsButton(context),
-    ]));
-  }
-
-  Widget _showTimeFrameDropdown(BuildContext context) {
-    return Expanded(
-        child: Padding(
-            padding: EdgeInsets.fromLTRB(
-                0.0, 0.0, UIConstants.SMALLER_PADDING / 2, UIConstants.SMALLER_PADDING / 2),
-            child: Container(
-                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
-                decoration: BoxDecoration(
-                    color: Theme.of(context).cardColor,
-                    borderRadius:
-                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
-                child: DropdownButtonHideUnderline(
-                    child: DropdownButton<String>(
-                  style: Theme.of(context).accentTextTheme.subtitle2,
-                  items: _createDropdownItems(TimeFrames.TIME_FRAMES.values.toList()),
-                  value: filterTimeframe,
-                  hint: Text("Time frame"),
-                  isExpanded: true,
-                  isDense: true,
-                  onChanged: (value) => setState(() {
-                    filterTimeframe = value;
-                    chartSeries = _buildChartSeries(context);
-                  }),
-                )))));
-  }
-
-  Widget _showLocationDropdown(BuildContext context) {
-    return Expanded(
-        child: Padding(
-            padding: EdgeInsets.fromLTRB(
-                UIConstants.SMALLER_PADDING / 2, 0.0, 0.0, UIConstants.SMALLER_PADDING / 2),
-            child: Container(
-                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
-                decoration: BoxDecoration(
-                    color: Theme.of(context).cardColor,
-                    borderRadius:
-                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
-                child: DropdownButtonHideUnderline(
-                    child: DropdownButton<String>(
-                  style: Theme.of(context).accentTextTheme.subtitle2,
-                  items: _createDropdownItems(widget.locationNamesToIds.keys.toList()),
-                  value: filterLocation,
-                  hint: Text("Location"),
-                  isExpanded: true,
-                  isDense: true,
-                  onChanged: (value) => setState(() {
-                    filterLocation = value;
-                    chartSeries = _buildChartSeries(context);
-                  }),
-                )))));
-  }
-
-  Widget _showSendTypeDropdown(BuildContext context) {
-    return Expanded(
-        child: Padding(
-            padding: EdgeInsets.fromLTRB(
-                0.0, UIConstants.SMALLER_PADDING / 2, UIConstants.SMALLER_PADDING / 2, 0.0),
-            child: Container(
-                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
-                decoration: BoxDecoration(
-                    color: Theme.of(context).cardColor,
-                    borderRadius:
-                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
-                child: DropdownButtonHideUnderline(
-                    child: DropdownButton<String>(
-                  style: Theme.of(context).accentTextTheme.subtitle2,
-                  items: _createDropdownItems(SendTypes.SEND_TYPES),
-                  value: filterSendType,
-                  hint: Text("Send type"),
-                  isExpanded: true,
-                  isDense: true,
-                  onChanged: (value) => setState(() {
-                    filterSendType = value;
-                    chartSeries = _buildChartSeries(context);
-                  }),
-                )))));
-  }
-
-  Widget _showCategoryDropdown(BuildContext context) {
-    return Expanded(
-        child: Padding(
-            padding: EdgeInsets.fromLTRB(
-                UIConstants.SMALLER_PADDING / 2, UIConstants.SMALLER_PADDING / 2, 0.0, 0.0),
-            child: Container(
-                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
-                decoration: BoxDecoration(
-                    color: Theme.of(context).cardColor,
-                    borderRadius:
-                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
-                child: DropdownButtonHideUnderline(
-                    child: DropdownButton<String>(
-                  style: Theme.of(context).accentTextTheme.subtitle2,
-                  items: _createDropdownItems(ClimbCategories.CATEGORIES),
-                  value: filterCategory,
-                  hint: Text("Category"),
-                  isExpanded: true,
-                  isDense: true,
-                  onChanged: (value) => setState(() {
-                    filterCategory = value;
-                    chartSeries = _buildChartSeries(context);
-                  }),
-                )))));
-  }
-
-  Widget _showClearDropdownsButton(BuildContext context) {
-    return Container(
-        child: IconButton(
-            icon: Icon(Icons.cancel,
-                color: (filterTimeframe == null &&
-                        filterLocation == null &&
-                        filterSendType == null &&
-                        filterCategory == null)
-                    ? Colors.grey
-                    : Theme.of(context).accentColor),
-            onPressed: () => setState(() {
-                  filterTimeframe = null;
-                  filterLocation = null;
-                  filterSendType = null;
-                  filterCategory = null;
-                  chartSeries = _buildChartSeries(context);
-                })));
-  }
-
-  List<DropdownMenuItem> _createDropdownItems(List<String> items) {
-    if (items.isNotEmpty) {
-      return items.map((String value) {
-        return DropdownMenuItem<String>(
-          value: value,
-          child: Text(value),
-        );
-      }).toList();
-    } else {
-      // null disables the dropdown
-      return null;
-    }
-  }
-
   Widget _buildChart(BuildContext context) {
-    Widget chart = (chartSeries != null)
-        ? charts.BarChart(chartSeries,
-            selectionModels: [
-              charts.SelectionModelConfig(
-                type: charts.SelectionModelType.info,
-                changedListener: _onSelectionChanged,
-              )
-            ],
-            primaryMeasureAxis: charts.NumericAxisSpec(
-                renderSpec: charts.GridlineRendererSpec(
-              labelStyle: charts.TextStyleSpec(
-                fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
-              ),
-              lineStyle: charts.LineStyleSpec(
-                color: charts.ColorUtil.fromDartColor(Theme.of(context).dialogBackgroundColor),
-              ),
-            )),
-            domainAxis: charts.OrdinalAxisSpec(
-              renderSpec: charts.SmallTickRendererSpec(
-                labelStyle: charts.TextStyleSpec(
-                  fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                  fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                  color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
-                ),
-              ),
-              tickProviderSpec: charts.StaticOrdinalTickProviderSpec(ticks),
-            ))
-        : Center(
-            child: Text(
-            "There are no existing attempts ${widget.attempts.isNotEmpty ? "matching these filters" : ""}. \nGo log some!",
-            style: Theme.of(context).accentTextTheme.subtitle2,
-            textAlign: TextAlign.center,
-          ));
     return Expanded(
         child: Padding(
             padding: EdgeInsets.only(top: UIConstants.SMALLER_PADDING),
@@ -275,12 +81,65 @@ class _AttemptsByTimeChartState extends State<AttemptsByTimeChart> {
                     color: Theme.of(context).cardColor,
                     borderRadius:
                         BorderRadius.all(Radius.circular(UIConstants.CARD_BORDER_RADIUS))),
-                child: chart)));
+                child: StreamBuilder(
+                    stream: filteredAttemptsStream.stream,
+                    initialData: widget.attempts,
+                    builder: (BuildContext context, snapshot) {
+                      List<charts.Series> chartSeries = _buildChartSeries(context, snapshot.data);
+                      return (chartSeries != null)
+                          ? charts.BarChart(chartSeries,
+                              selectionModels: [
+                                charts.SelectionModelConfig(
+                                  type: charts.SelectionModelType.info,
+                                  changedListener: _onSelectionChanged,
+                                )
+                              ],
+                              primaryMeasureAxis: charts.NumericAxisSpec(
+                                  renderSpec: charts.GridlineRendererSpec(
+                                labelStyle: charts.TextStyleSpec(
+                                  fontSize:
+                                      Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
+                                  fontWeight: Theme.of(context)
+                                      .accentTextTheme
+                                      .caption
+                                      .fontWeight
+                                      .toString(),
+                                  color:
+                                      charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                                ),
+                                lineStyle: charts.LineStyleSpec(
+                                  color: charts.ColorUtil.fromDartColor(
+                                      Theme.of(context).dialogBackgroundColor),
+                                ),
+                              )),
+                              domainAxis: charts.OrdinalAxisSpec(
+                                renderSpec: charts.SmallTickRendererSpec(
+                                  labelStyle: charts.TextStyleSpec(
+                                    fontSize:
+                                        Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
+                                    fontWeight: Theme.of(context)
+                                        .accentTextTheme
+                                        .caption
+                                        .fontWeight
+                                        .toString(),
+                                    color: charts.ColorUtil.fromDartColor(
+                                        Theme.of(context).accentColor),
+                                  ),
+                                ),
+                                tickProviderSpec: charts.StaticOrdinalTickProviderSpec(ticks),
+                              ))
+                          : Center(
+                              child: Text(
+                              "There are no existing attempts ${widget.attempts.isNotEmpty ? "matching these filters" : ""}. \nGo log some!",
+                              style: Theme.of(context).accentTextTheme.subtitle2,
+                              textAlign: TextAlign.center,
+                            ));
+                    }))));
   }
 
-  List<charts.Series<AttemptsByTimeSeries, String>> _buildChartSeries(BuildContext context) {
+  List<charts.Series<AttemptsByTimeSeries, String>> _buildChartSeries(
+      BuildContext context, List<Attempt> filteredAttempts) {
     _resetChartSelection();
-    List<Attempt> filteredAttempts = _filterAttempts();
 
     if (filteredAttempts.isEmpty) {
       return null;
@@ -332,49 +191,12 @@ class _AttemptsByTimeChartState extends State<AttemptsByTimeChart> {
   }
 
   void _resetChartSelection() {
-    setState(() {
-      selectedHour = null;
-      selectedCount = null;
-    });
-  }
-
-  List<Attempt> _filterAttempts() {
-    List<Attempt> filteredAttempts = widget.attempts;
-
-    if (filterTimeframe != null) {
-      if (filterTimeframe == TimeFrames.TIME_FRAMES["lastWeek"]) {
-        filteredAttempts = filteredAttempts
-            .where((attempt) =>
-                DateTime.now().difference(attempt.timestamp.toDate()) < Duration(days: 7))
-            .toList();
-      } else if (filterTimeframe == TimeFrames.TIME_FRAMES["lastMonth"]) {
-        filteredAttempts = filteredAttempts
-            .where((attempt) =>
-                DateTime.now().difference(attempt.timestamp.toDate()) < Duration(days: 30))
-            .toList();
-      } else if (filterTimeframe == TimeFrames.TIME_FRAMES["lastYear"]) {
-        filteredAttempts = filteredAttempts
-            .where((attempt) =>
-                DateTime.now().difference(attempt.timestamp.toDate()) < Duration(days: 365))
-            .toList();
-      }
+    if (selectedHour != null) {
+      setState(() {
+        selectedHour = null;
+        selectedCount = null;
+      });
     }
-    if (filterLocation != null) {
-      filteredAttempts = filteredAttempts
-          .where((attempt) => attempt.locationId == widget.locationNamesToIds[filterLocation])
-          .toList();
-    }
-    if (filterSendType != null) {
-      filteredAttempts =
-          filteredAttempts.where((attempt) => attempt.sendType == filterSendType).toList();
-    }
-    if (filterCategory != null) {
-      filteredAttempts = filteredAttempts
-          .where((attempt) => attempt.climbCategories.contains(filterCategory))
-          .toList();
-    }
-
-    return filteredAttempts;
   }
 
   _onSelectionChanged(charts.SelectionModel model) {
@@ -392,6 +214,12 @@ class _AttemptsByTimeChartState extends State<AttemptsByTimeChart> {
       selectedHour = hour;
       selectedCount = count;
     });
+  }
+
+  @override
+  void dispose() {
+    filteredAttemptsStream.close();
+    super.dispose();
   }
 }
